add click damage to upgrade system

+++ Assets/Code/Scripts/MVC/Models/SettingsModel.cs
@@ -33,7 +33,7 @@ public class SettingsModel : MonoBehaviour
     public bool changeTimeScale;
     [ConditionalField(nameof(changeTimeScale))]
     public float timeScale = 1f;
-    public float clickDamage;
+    public double clickDamage;
 
     [Header("Settings edited from game UI")]
     #region Is60fps
